fix(web): resolve circular dependency between useChat and sub-engines
